Refactored common MPI-related setup/teardown code into a base class, updated comments, added FIXMEs.
