+++ cpp/tests/utilities/mg_test_utilities.hpp
@@ -17,6 +17,10 @@
 
 #include <experimental/graph.hpp>
 
+#include <raft/comms/mpi_comms.hpp>
+
+#include <gtest/gtest.h>
+
 namespace cugraph {
 namespace test {
 
@@ -31,6 +35,43 @@ std::tuple<
                     graph_t<vertex_t, edge_t, weight_t, store_transposed, true>>,  // multi_gpu=true
   rmm::device_uvector<vertex_t>>
 create_graph_for_gpu(raft::handle_t& handle, const std::string& graph_file_path);
+
+/**
+ * @brief Base test fixture class, responsible for handling common operations
+ * needed by all MG tests.
+ *
+ * It's expected this class will be built out and refactored often as new MG C++
+ * tests are added and new patterns evolve.
+ *
+ * Example:
+ * ```
+ * class MyTestFixture : public cugraph::test::MG_TestFixture_t<My_Testparams_t> {};
+ * ```
+ **/
+
+// FIXME: consider moving this to a separate file? (eg. mg_test_fixture.cpp)?
+
+class MG_TestFixture_t : public ::testing::Test {
+ public:
+  static void SetUpTestCase()
+  {
+    MPI_TRY(MPI_Init(NULL, NULL));
+
+    int rank, size;
+    MPI_TRY(MPI_Comm_rank(MPI_COMM_WORLD, &rank));
+    MPI_TRY(MPI_Comm_size(MPI_COMM_WORLD, &size));
+
+    int nGpus;
+    CUDA_CHECK(cudaGetDeviceCount(&nGpus));
+
+    ASSERT(
+      nGpus >= size, "Number of GPUs are lesser than MPI ranks! ngpus=%d, nranks=%d", nGpus, size);
+
+    CUDA_CHECK(cudaSetDevice(rank));
+  }
+
+  static void TearDownTestCase() { MPI_TRY(MPI_Finalize()); }
+};
 
 }  // namespace test
 }  // namespace cugraph
